Research_Assistant/requirements.txt
# Core dependencies
streamlit>=1.22.0
python-dotenv>=1.0.0
requests>=2.28.2

# PDF processing
PyMuPDF>=1.22.3

# Embedding and vector search
sentence-transformers>=2.2.2
chromadb>=0.4.6

# Text processing and analysis
scikit-learn>=1.2.2
numpy>=1.24.3

# Advanced retrieval
transformers>=4.30.2

# Monitoring and logging
uuid>=1.30 